Refactor service
