+++ starter/critter/src/main/java/com/udacity/jdnd/course3/critter/services/impl/ScheduleServiceImpl.java
@@ -1,0 +1,94 @@
+package com.udacity.jdnd.course3.critter.services.impl;
+
+import com.udacity.jdnd.course3.critter.entities.CustomerEntity;
+import com.udacity.jdnd.course3.critter.entities.EmployeeEntity;
+import com.udacity.jdnd.course3.critter.entities.PetEntity;
+import com.udacity.jdnd.course3.critter.entities.ScheduleEntity;
+import com.udacity.jdnd.course3.critter.repositories.CustomerRepository;
+import com.udacity.jdnd.course3.critter.repositories.EmployeeRepository;
+import com.udacity.jdnd.course3.critter.repositories.PetRepository;
+import com.udacity.jdnd.course3.critter.repositories.ScheduleRepository;
+import com.udacity.jdnd.course3.critter.services.ScheduleService;
+import org.springframework.stereotype.Service;
+
+import javax.persistence.EntityNotFoundException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Optional;
+
+@Service
+public class ScheduleServiceImpl implements ScheduleService {
+
+    private final ScheduleRepository scheduleRepository;
+    private final PetRepository petRepository;
+    private final EmployeeRepository employeeRepository;
+    private final CustomerRepository customerRepository;
+
+    public ScheduleServiceImpl(ScheduleRepository scheduleRepository, PetRepository petRepository, EmployeeRepository employeeRepository, CustomerRepository customerRepository) {
+        this.scheduleRepository = scheduleRepository;
+        this.petRepository = petRepository;
+        this.employeeRepository = employeeRepository;
+        this.customerRepository = customerRepository;
+    }
+
+    /**
+     * Create schedule. Get corresponding pet entities from the pet ids and also employee entities from employee ids to be save into the schedule
+     *
+     * @param scheduleEntity to be saved
+     * @param petIds         list of pet ids
+     * @param employeeIds    list of employee ids
+     * @return saved schedule
+     */
+    @Override
+    public ScheduleEntity createSchedule(ScheduleEntity scheduleEntity, List<Long> petIds, List<Long> employeeIds) {
+        List<PetEntity> petEntities = new ArrayList<>();
+        if (petIds != null && !petIds.isEmpty()) {
+            petIds.forEach(petId -> {
+                Optional<PetEntity> petEntityOptional = petRepository.findById(petId);
+                petEntityOptional.ifPresent(petEntities::add);
+            });
+        }
+        List<EmployeeEntity> employeeEntities = new ArrayList<>();
+        if (employeeIds != null && !employeeIds.isEmpty()) {
+            employeeIds.forEach(employeeId -> {
+                Optional<EmployeeEntity> employeeEntityOptional = employeeRepository.findById(employeeId);
+                employeeEntityOptional.ifPresent(employeeEntities::add);
+            });
+        }
+
+
+        scheduleEntity.setPets(petEntities);
+        scheduleEntity.setEmployees(employeeEntities);
+
+        return scheduleRepository.save(scheduleEntity);
+    }
+
+    @Override
+    public List<ScheduleEntity> getAllSchedules() {
+        return scheduleRepository.findAll();
+    }
+
+    @Override
+    public List<ScheduleEntity> getScheduleForPet(long petId) {
+        return scheduleRepository.findAllByPetsId(petId);
+    }
+
+    @Override
+    public List<ScheduleEntity> getScheduleForEmployee(long employeeId) {
+        return scheduleRepository.findAllByEmployeesId(employeeId);
+    }
+
+    @Override
+    public List<ScheduleEntity> getScheduleForCustomer(long customerId) {
+        Optional<CustomerEntity> customerEntityOptional = customerRepository.findById(customerId);
+
+        CustomerEntity customerEntity = customerEntityOptional.orElseThrow(EntityNotFoundException::new);
+        List<PetEntity> petEntityList = customerEntity.getPets();
+        List<ScheduleEntity> scheduleEntities = new ArrayList<>();
+
+        petEntityList.forEach(petEntity -> {
+            scheduleEntities.addAll(scheduleRepository.findAllByPetsId(petEntity.getId()));
+        });
+        return scheduleEntities;
+    }
+}
